# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "summary": "Login User",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User logged out successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Logout",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset link sent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "summary": "Forgot password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "summary": "Reset password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/verify": {
      "get": {
        "operationId": "AuthController_verifyEmail",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "description": "Verification token from email",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email verified successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Invalid or expired token.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Verify user email via token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/resend-verification": {
      "post": {
        "operationId": "AuthController_resendVerification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendVerificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification email resent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Bad request (e.g. user not found or already verified).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          }
        },
        "summary": "Resend email verification link",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokensResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/billing/plans": {
      "get": {
        "operationId": "BillingController_listPlans",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "List available plans",
        "tags": [
          "Billing"
        ]
      }
    },
    "/billing/subscribe": {
      "post": {
        "operationId": "BillingController_subscribe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscribeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Create or switch subscription",
        "tags": [
          "Billing"
        ]
      }
    },
    "/billing/cancel": {
      "delete": {
        "operationId": "BillingController_cancel",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Cancel current subscription",
        "tags": [
          "Billing"
        ]
      }
    },
    "/billing/invoices": {
      "get": {
        "operationId": "BillingController_invoices",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "List invoices for current user",
        "tags": [
          "Billing"
        ]
      }
    },
    "/billing/callback": {
      "post": {
        "operationId": "BillingController_callback",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Payment gateway callback",
        "tags": [
          "Billing"
        ]
      }
    },
...
  },
  "info": {
    "title": "YSHAI API",
    "description": "Generated for Orval client",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "user@example.com",
            "maxLength": 50,
            "minLength": 5,
            "format": "email"
          },
          "password": {
            "type": "string",
            "description": "Password for authentication",
            "example": "strongPass123",
            "minLength": 6,
            "maxLength": 100,
            "format": "password"
          },
          "name": {
            "type": "string",
            "description": "Full name of the user",
            "example": "John Doe",
            "minLength": 2,
            "maxLength": 100
          },
          "timezone": {
            "type": "string",
            "description": "Timezone of the user",
            "example": "America/New_York"
          },
          "timeFormat": {
            "type": "string",
            "description": "Preferred time format of the user",
            "example": "12h"
          }
        },
        "required": [
          "email",
          "password",
          "name",
          "timezone",
          "timeFormat"
        ]
      },
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Operation successful"
          }
        },
        "required": [
          "message"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "format": "email",
            "minLength": 5,
            "maxLength": 50
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Password reset token"
          },
          "newPassword": {
            "type": "string",
            "minLength": 6,
            "description": "New password"
          }
        },
        "required": [
          "token",
          "newPassword"
        ]
      },
      "ResendVerificationDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email to resend verification link to",
            "example": "user@example.com",
            "format": "email"
          }
        },
        "required": [
          "email"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          },
          "userId": {
            "type": "number"
          }
        },
        "required": [
          "refreshToken",
          "userId"
        ]
      },
      "TokensResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ]
      },
...
      "PlanResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string",
            "enum": [
              "Free",
              "Pro",
              "Business"
            ]
          },
          "priceMonthly": {
            "type": "number"
          },
          "maxAccounts": {
            "type": "number"
          },
          "aiCreditsUnlimited": {
            "type": "boolean"
          },
          "aiCreditsLimit": {
            "type": "object",
            "nullable": true
          },
          "teamCollaboration": {
            "type": "boolean"
          },
          "analyticsExport": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "priceMonthly",
          "maxAccounts",
          "aiCreditsUnlimited",
          "teamCollaboration",
          "analyticsExport",
          "createdAt"
        ]
      },
      "SubscribeDto": {
        "type": "object",
        "properties": {
          "planId": {
            "type": "number",
            "description": "Plan ID to subscribe to"
          },
          "paymentMethod": {
            "type": "string",
            "description": "Payment method token or identifier"
          },
          "couponCode": {
            "type": "string",
            "description": "Optional coupon code"
          }
        },
        "required": [
          "planId",
          "paymentMethod"
        ]
      },
      "SubscriptionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "canceled",
              "past_due",
              "trialing"
            ]
          },
          "periodStartsAt": {
            "format": "date-time",
            "type": "string"
          },
          "periodEndsAt": {
            "format": "date-time",
            "type": "string"
          },
          "canceledAt": {
            "type": "object",
            "nullable": true
          },
          "paymentGatewaySubscriptionId": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "planId": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "status",
          "periodStartsAt",
          "periodEndsAt",
          "createdAt",
          "updatedAt",
          "planId"
        ]
      },
      "SubscribeResponseDto": {
        "type": "object",
        "properties": {
          "subscription": {
            "$ref": "#/components/schemas/SubscriptionResponseDto"
          },
          "clientSecret": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "subscription"
        ]
      },
      "InvoiceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string",
            "enum": [
              "SAR",
              "USD",
              "EUR"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "paid",
              "unpaid",
              "refunded",
              "failed"
            ]
          },
          "paymentGatewayId": {
            "type": "object",
            "nullable": true
          },
          "paymentMethod": {
            "type": "object",
            "nullable": true
          },
          "issuedAt": {
            "format": "date-time",
            "type": "string"
          },
          "paidAt": {
            "type": "object",
            "nullable": true
          },
          "downloadedAt": {
            "type": "object",
            "nullable": true
          },
          "pdfUrl": {
            "type": "object",
            "nullable": true
          },
          "metadata": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "amount",
          "currency",
          "status",
          "issuedAt"
        ]
      },
...
    }
  }
}